Working-ish version of the jank filters.
SimonIIR now runs an interrupt at 4 kHz. It looks at an in band
lowpass and an out of band highpass. When the normalized energy
average of the low pass filter is greater than the high pass filter,
then sample flag is pulled high, and the LED turns on.

The interrupt only works when sample_flag is low, and will resume
checking when sample_flag is high.